evaluation/competency_questions/requirements.txt
rdflib
pyyaml